% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{online}
\alias{online}
\title{Check if the host of a URL is online}
\usage{
online(url = "google.com")
}
\arguments{
\item{url}{a URL to check}
}
\value{
boolean
}
\description{
Check if the host of a URL is online
}
\examples{
online()
}
