% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid.R
\name{orcid_person}
\alias{orcid_person}
\title{Get Person Details for ORCiD}
\usage{
orcid_person(orcid)
}
\arguments{
\item{orcid}{a vector of ORCiDs}
}
\value{
A data frame of details
}
\description{
Get Person Details for ORCiD
}
\examples{
\donttest{
orcids <- c("0000-0002-0247-239X", "0000-0002-7523-5539")
orcid_person(orcids)
}
}
