% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep_if}
\alias{rep_if}
\title{Replace If}
\usage{
rep_if(x, y, replace = NULL)
}
\arguments{
\item{x}{For each \code{x[[i]]} if a value in \code{replace}, return \code{y[[i]]}; otherwise return \code{x[[i]]}.}

\item{y}{Replacement value(s); if not the same length as \code{x}, then values will be recycled}

\item{replace}{values in \code{x} to replace}
}
\description{
Replace values if NULL, NA, or specified value
}
\examples{
rep_if(list(NULL, 1), 2)
rep_if(list(NULL, 0, NULL), 1:3)
}
\keyword{internal}
