% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verbose}
\alias{verbose}
\title{Set or get verbosity}
\usage{
verbose(verbose = NULL)
}
\arguments{
\item{verbose}{if logical, sets whether to show verbose output messages and progress bars}
}
\value{
the current option value (logical)
}
\description{
Set or get verbosity
}
\examples{
verbose()
}
