% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{xml_find}
\alias{xml_find}
\title{Find and return info from XML by xpath}
\usage{
xml_find(xml, xpath, join = NULL)
}
\arguments{
\item{xml}{the xml document, node, or nodeset}

\item{xpath}{a string containing an xpath expression}

\item{join}{optional string to join vectors}
}
\value{
text
}
\description{
Find and return info from XML by xpath
}
